ref!: pipe all logging to stderr
Sending all logs to stderr makes linode-tagger's output much simplier to
pipe to jq from `--json`, or to redirect the `--report` to a file.

I also considered only sending Info/Debug to stderr if report/json flags
were used, but figured it was a worthwhile change regardless. I'm open
to changing this though!