ref!: use discrete structs rather than raw interfaces/maps
This commit is intended to make the code both easier to read/follow, as
well as prevent accidental logic errors that might come about from using
raw interfaces and composite maps.

This commit is a fairly large refactor, as it seemed it would be even
more complicated to try and get functional code in partial commits to
try and make the changeset easier to follow. To that end, this commit is
add-only; it does not remove any old code, it simply adds new functions
and updates the call chain in main.

Changes included:
- new structs to hold organized data for diffs and "desired" tag data
(ie, data that will be used to update objects on the account)
- new composite structs (collections) that hold diffs/desired tag data
sets for each object type (instances, volumes, domains, nodebalancers,
lke clusters)
- new diff/comparison functions that work with the new structs
- fixed/improved diff checking when updating tags (ie, not using
`dry-run`)

Still to be done:
- `genReport()` func needs to be updated to use new structs
- clean/remove old code