fix: don't include unconfigured objects in desired tag changes/diffs
This commit updates the `getNewTags()` func to also return a boolean
value that indicates whether or not the given object label has a tag
rule configured for it. If it does not match a configured tag rule, we
should not touch the resource.

This also adds a bit of extra logic to ensure that objects that _do_
match a tag rule but _do not_ have any requested changes don't make it
into the desired tag changes/diffs, since there are no changes.